Statistical types and typesets: requires refactoring of testing structure for multiple typesets.
