+++ src/visions/types/qualitative_quantitative/__init__.py
@@ -1,0 +1,6 @@
+from visions.types.qualitative_quantitative.continuous import Continuous
+from visions.types.qualitative_quantitative.discrete import Discrete
+from visions.types.qualitative_quantitative.nominal import Nominal
+from visions.types.qualitative_quantitative.ordinal import Ordinal
+from visions.types.qualitative_quantitative.qualitative import Qualitative
+from visions.types.qualitative_quantitative.quantitative import Quantitative
